requirements.txt
aioredis==1.1.0
appnope==0.1.0
asgiref==3.2.3
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
autobahn==19.3.3
Automat==0.7.0
awscli==1.16.9
awsebcli==3.14.4
blessed==1.15.0
boto3==1.9.203
botocore==1.11.9
CacheControl==0.12.5
cached-property==1.4.3
cachetools==3.1.0
cement==2.8.2
certifi==2018.1.18
cffi==1.12.2
channels-redis==2.2.1
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.6
coreschema==0.0.4
cryptography==2.6.1
daphne==2.2.2
decorator==4.1.2
Django==3.0.2
django-appconf==1.0.3
django-braces==1.12.0
django-cleanup==4.0.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-resized==0.3.9
django-rest-knox==4.1.0
django-sslserver==0.20
django-storages==1.6.6
django-webpack-loader==0.6.0
django-websocket-redis==0.5.2
djangorestframework==3.11.0
docker==3.5.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
ffmpeg-python==0.2.0
firebase-admin==3.2.1
fsspec==0.4.1
future==0.17.1
gevent==1.3.6
google-api-core==1.15.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.1.0
google-cloud-firestore==1.6.1
google-cloud-storage==1.24.1
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
greenlet==0.4.14
grpcio==1.19.0
hiredis==0.2.0
httpie==0.9.9
httplib2==0.15.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipython==6.1.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.10.2
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
Markdown==2.6.11
MarkupSafe==1.0
msgpack==0.5.6
pathspec==0.5.5
pexpect==4.2.1
pickleshare==0.7.4
Pillow==6.1.0
prompt-toolkit==1.0.14
protobuf==3.7.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
Pygments==2.2.0
PyHamcrest==1.9.0
python-dateutil==2.7.3
pytz==2019.3
PyYAML==3.13
redis==2.10.6
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
semantic-version==2.5.0
service-identity==17.0.0
shutilwhich==1.1.0
simplegeneric==0.8.1
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
traitlets==4.3.2
Twisted==18.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.17.1
virtualenv==15.1.0
wcwidth==0.1.7
websocket-client==0.49.0
zope.interface==4.5.0
